Change length of xtime variables to ShortStrKIND to match netCDF files
This commit changes the declared length of xtime and similar timestamp
variables in the mpas_io_streams module from StrKIND to ShortStrKIND to
match the length of xtime in netCDF files.

Prior to this change, because the in-memory xtime variable was larger
than the in-file version, reading xtime could result in a string with
garbage characters at the end (beginning at character ShortStrKIND+1).
These gargbage characters, in turn, could lead to errors in the timekeeping
module when trying to parse the string as a timestamp. By matching
the in-memory size of the xtime variable with its size in files,
we avoid this problem.