ISO_C_BINDING update for pool hash
This commit creates an `ISO_C_BINDING` interface for the `pool_hash.c` file. To
handle the interface between Fortran and C, a new function (`pool_hash(...)`
was created in `mpas_pool_routines.F` to act as a wrapper for the `pool_hash.c`
function (`c_pool_hash(...)).

The wrapper function converts the Fortran types into corresponding C types and
calls `c_pool_hash`.

The `len` argument for `c_pool_hash` has been removed and has been
replaced with `strlen(key)` accordingly. Note that this works now because of
the use of `mpas_f_to_c_string` function module, which prepends a null
character onto `key`.

As well note, that if it is desired to create a hash that is greater then
`strlen(key)`, the length will need to be re-added in another commit.

Also contains the fix (from the PR review) to loop until the null character
within `key` instead of using `strlen`.