Avoid use of paths in include statements for registry-generated files in MPAS-A
In both the atmosphere and init_atmosphere core interface modules, files
generated by the registry were assumed to be found in a sub-directory named
"inc/". When building MPAS-A in alternate ways, the "*.inc" files generated
by the registry program may be located elsewhere.

This commit removes the "inc/" prefix for *.inc files included in the core
interface modules, which necessitates the addition of -I./inc in the Makefiles
for the atmosphere and init_atmosphere cores.