2 additional upper absorbing layer options are added:
(1) The absorbing layer often used in the MPAS/CAM configuration.  This option
activates a fixed second-order horizontal diffusion in the topmost
three layers on the dynamics variables (u, w and theta).
(2) Rayleigh damping on the horizontal mumentum (u) with a user-specified
damping timescale and a linear ramp over a number of levels from the top
(also user specified).