WSM6 correction for accretion for freezing rain
Prior to these modifications, this scheme would allow rain to freeze too easily,
with even very small amounts of ice. The changes reduce the efficiency
proportional to the sqaure of the ratio of ice to water during the accretion
process. This maintains liquid precipitation longer, which is in better
agreement with observations.