Update checkout_data_files.sh to obtain new physics tables for WRF 4.0.3 physics
The new checkout_data_files.sh looks for physics lookup tables that are
compatible with MPAS-Atmosphere version 7.0. In the new checkout script, we look
for a tag in the MPAS-Data repository that matches the version of the physics
files matching the MPAS-Atmosphere version.

The plan is that, moving forward, new editions of physics lookup tables will
always be found in a tag named according to the version of the tables, and
the 'master' branch in the MPAS-Data repository will essentially remain
unchanged since the version 3.0 physics tables to maintain compatibility with
older versions of MPAS-Atmosphere, whose checkout script only looks in
the master branch of the repository.