Don't process trigraphs with the XL compilers
The XL Fortran compiler invokes the XL C preprocessor, which by default processes
trigraphs (https://en.wikipedia.org/wiki/Digraphs_and_trigraphs#C). This can
cause problems in MPAS code, e.g., if a string '??' appears in a non-comment
context in the source, where the ??' characters are turned into a ^ character,
leading to an un-terminated character string.

To avoid processing of trigraphs, the -qnotrigraph option is passed to the C
preprocessor via -WF in the FFLAGS for the XL compiler only. There is apparently
no need to add -qnotrigraph to the CPPFLAGS, since the 'cpp' that is invoked
when explicitly preprocessing files with GEN_F90=true is the system's
C preprocessor, which doesn't process trigraphs.