Implement supersampling of albedo fields for high-res MPAS meshes
The MODIS monthly albedo and maximum snow albedo fields are provided at
0.05-degree resolution. When the grid cells in an MPAS mesh are smaller than
this, there will in general be MPAS cells to which we don't map a source
data value.

To work around this issue, this commit adds the capability to supersample
the MODIS monthly albedo and maximum surface albedo datasets by some integer
factor. This supersampling factor is given by a new namelist option,
config_supersample_factor, available in the &data_sources namelist record.
The default supersampling factor is 1.