Avoid potential divide-by-zero in mpas_atm_update_bdy_tend
When computing rho_zz, we use array syntax to divide rho(:,:) by zz(:,:).
The value of zz in the "garbage cell" is generally not predictable, and
this can lead to a divide by zero FPE. To avoid this potential FPE, we
now set zz(:,nCells+1) = 1.0 in mpas_atm_update_bdy_tend. In principle
this should be completely safe, since no other code should rely on
the garbage element for a field taking on any particular value.