Verify that config_apply_lbcs is consistent with values in bdyMaskCell field
This commit adds checks in the mpas_atm_bdy_checks routine to ensure that
the value of config_apply_lbcs is consistent with the presence of boundary
elements in the bdyMaskCellField.

When config_apply_lbcs = false, the bdyMaskCell field should be zero everywhere,
indicating no boundary elements. Also, when config_apply_lbcs = true, there must
be positive values in the bdyMaskCell field.