Enable soundings to handle non-existence of config_sounding_interval option
When MPAS-Atmosphere is built as a dycore in other modeling systems
(e.g, CAM/CESM), only a subset of the stand-alone MPAS-Atmosphere namelist
options may be available. To avoid segfaults in the soundings diagnostic
module due to the non-existence of the config_sounding_interval option in
such cases, add extra logic in the soundings_setup routine to detect that
this namelist option is not available and deactivate soundings just as we
would if config_soundings_interval = 'none'.