Remove broken log write of cf1,cf2,cf3 from init_atm_case_gfs
The init_atm_case_gfs routine previously had a call to mpas_log_write
that was intended to print out the values of cf1, cf2, and cf3. However,
this call was missing '$r' conversion specifications for the cf{1,2,3}
variables.

Rather than fixing this mpas_log_write call, it seems cleaner to simply
remove it, since the values of cf1, cf2, and cf3 do not appear to be used
directly in the init_atm_case_gfs routine, and the values of these variables
can be found in the model initial conditions file.