Fix race condition in file tests when clobber_mode = 'append' or 'overwrite'
The stream manager's write_stream(...) routine tests on the existence of files
to be written when clobber_mode = 'append' or 'overwrite' for a stream.
Each MPI task was independently testing whether the output file existed using
calls to Fortran's inquire intrinsic, and the results of these inquiries could
disagree among MPI tasks, apparently due to race conditions between the creation
of output files and the testing of files on different tasks.

The solution adopted in this commit is essentially the same approach as is used
in the low-level mpas_io_open(...) routine: we test for the existence of the file
only on the IO_NODE and broadcast the result to all other MPI tasks.

There are two places in write_stream(...) where file inquiries are modified.
There is also one file inquiry in read_stream(...) that has been modified, too,
not out of concerns for correctness, but simply because it seemed more efficient
to make filesystem inquiries only from one MPI task and to broadcast the result
than to have a potentially large number of MPI tasks all making filesystem
inquiries.