Return error from read_stream() when inexact stream time not found
The MPAS_stream_mgr_read() routine calls read_stream() to handle the actual
reading of an individual stream. Previously, the read_stream routine did not
return an error code if the requested stream time was not MPAS_STREAM_EXACT_TIME
and the requested time could not be found. For example, if the earliest time
strictly after the current time was requested from a stream, but no file with
data after the current time was found, the read_stream routine would not return
an error, but would incorrectly try to read from record 0 (which is not valid)
from whatever file was last opened by the stream.

The fix in this commit simply sets the error code to MPAS_STREAM_MGR_ERROR
and returns when the Fortran INQUIRE statement for the appropriate file returns
false (indicating that the needed file is not found).

Note that in the active code block where the error code is now returned, we need
to look for (i.e., "retest") a file because we are searching for an inexact
time. If the file is not found, then retestFile is set to .false., and the code
block beginning at what is now L.3829 in mpas_stream_manager.F will not be
executed; accordingly, rebuildStream will never be set to .true., and the code
block beginning at what is now L.3880 will not be executed, in which case
the call to MPAS_readStream at L.3926 will not perform a valid read for
the requested stream time.