Fix weighting in ALE SSH thickness distribution
In eqn 6 of Petersen et al 2015, remove h_k^{rest}. This allows the
resting thickness to be zero, which can occur with wetting/drying and
sigma coordinates.

For the initial condition, require that
sum (restingThickness ) = bottomDepth for each column.
This is always done for z-star global ICs, but should be done for sigma
as well.